requirements.txt
pymongo>=4.0
python-dotenv>=1.0
requests>=2.28
